slack-bot/requirements.txt
slack-bolt==1.18.0
requests==2.31.0
python-dotenv==1.0.0
apscheduler==3.10.4
